bcrypt + module cleanup
