+++ Package.swift
@@ -2,16 +2,11 @@ import PackageDescription
 
 let package = Package(
     name: "Crypto",
-    targets: [
-        Target(name: "Essentials"),
-        Target(name: "BCrypt", dependencies: ["Random"]),
-        Target(name: "Hash", dependencies: ["Essentials", "Random"]),
-        Target(name: "Random", dependencies: ["Essentials"]),
-        Target(name: "HMAC", dependencies: ["Essentials", "Random"]),
-        Target(name: "Cipher", dependencies: ["Essentials"]),
-    ],
     dependencies: [
-        .Package(url: "https://github.com/vapor/core.git", Version(2,0,0, prereleaseIdentifiers: ["alpha"])),
+        // Module for generating random bytes and numbers.
+        .Package(url: "https://github.com/vapor/random.git", majorVersion: 0),
+
+        // LibreSSL wrapped in a Swift package.
         .Package(url: "https://github.com/vapor/clibressl.git", majorVersion: 1)
     ]
 )
