+++ Sources/Crypto/Cipher/Cipher.swift
@@ -1,6 +1,4 @@
 import CLibreSSL
-import Essentials
-import Core
 
 public final class Cipher {
     public let method: Method
