+++ Sources/Crypto/Essentials/ByteStream.swift
