+++ Sources/Crypto/Essentials/Exports.swift
@@ -1,0 +1,2 @@
+@_exported import Core
+@_exported import Random
