+++ Sources/Crypto/HMAC/HMAC+Method.swift
@@ -1,5 +1,3 @@
-import Core
-import Essentials
 import CLibreSSL
 
 /// Hashing method for calculating
