+++ Sources/Crypto/Hash/Hash+Method.swift
