+++ Tests/LinuxMain.swift
@@ -1,19 +1,8 @@
 import XCTest
-@testable import CipherTests
-@testable import HashTests
-@testable import HMACTests
-@testable import RandomTests
+@testable import CryptoTests
 
 XCTMain([
-    // Cipher
     testCase(CipherTests.allTests),
-
-    // Hash 
     testCase(HashTests.allTests),
-    
-    // HMAC
     testCase(HMACTests.allTests),
-
-    // Random
-    testCase(RandomTests.allTests),
 ])
