+++ circle.yml
@@ -1,3 +1,3 @@
 test:
   override:
-    - eval "$(curl -sL swift.vapor.sh/ci)"
+    - eval "$(curl -sL swift.vapor.sh/ci-3.1)"
